requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
dotenv==0.9.9
idna==3.11
joblib==1.5.2
mysql-connector-python==9.5.0
mysqlclient==2.2.7
nltk==3.9.2
numpy==2.2.6
python-dotenv==1.2.1
regex==2025.11.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
threadpoolctl==3.6.0
tqdm==4.67.1
urllib3==2.5.0
